Support bucket load in other contexts than ScriptRunner
